+++ riscv/src/lib.rs
@@ -40,6 +40,10 @@ pub(crate) mod bits;
 pub mod delay;
 pub mod interrupt;
 pub mod register;
+
+// Re-export crates of the RISC-V ecosystem
+#[cfg(feature = "riscv-macros")]
+pub use riscv_macros::*;
 pub use riscv_pac::*;
 
 #[macro_use]