+++ tests/tests/riscv-rt/external_interrupt/fail_impl_interrupt_number.rs
@@ -1,0 +1,4 @@
+#[riscv_rt::external_interrupt(riscv::interrupt::Interrupt::SupervisorSoft)]
+fn my_interrupt() {}
+
+fn main() {}