+++ riscv-semihosting/CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- Apply clippy changes
+- Bump riscv dependency version
 - Made `cfg` variable selection more robust for custom targets
 - Fixed debug::exit() on riscv64 QEMU simulation
 - Fixed an ambiguous link in the generated crate documentation.